Fix test: moved to other test which cleanup after-method
